# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server CRM server.",
        "title": "Golang CRM Swagger Documentation",
        "contact": {
            "name": "API Support",
            "url": "https://t.me/xalmatoff",
            "email": "venom.uzz@mail.ru"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/v1/groups": {
            "post": {
                "description": "This API to create a group.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Groups"
                ],
                "summary": "Create a group.",
                "parameters": [
                    {
                        "type": "string",
                        "x-order": "1",
                        "name": "url",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "x-order": "10",
                        "name": "seoDescription",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "x-order": "11",
                        "name": "seoKeywords",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "x-order": "12",
                        "name": "seoText",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "x-order": "13",
                        "name": "seoTitle",
                        "in": "formData"
                    },
                    {
                        "type": "boolean",
                        "x-order": "14",
                        "name": "enabled",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "x-order": "2",
                        "name": "nameUz",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "x-order": "3",
                        "name": "nameEn",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "x-order": "3",
                        "name": "nameRu",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "x-order": "4",
                        "name": "descriptionUz",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "x-order": "5",
                        "name": "descriptionRu",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "x-order": "6",
                        "name": "descriptionEn",
                        "in": "formData"
                    },
                    {
                        "type": "integer",
                        "x-order": "7",
                        "name": "position",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "x-order": "8",
                        "name": "parentGroup",
                        "in": "formData"
                    },
                    {
                        "type": "file",
                        "description": "fileImage",
                        "name": "fileImage",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Groups"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/v1.Response"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/v1.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/v1.Response"
                        }
                    }
                }
            }
        },
        "/v1/groups-by-filter": {
            "get": {
                "description": "This API to get groups by filter.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Groups"
                ],
                "summary": "Get groups by filter.",
                "parameters": [
                    {
                        "type": "boolean",
                        "x-order": "1",
                        "name": "all",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "x-order": "2",
                        "name": "parentId",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "x-order": "3",
                        "name": "nameUz",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "x-order": "3",
                        "name": "url",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "x-order": "4",
                        "name": "enabled",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.GroupsWithChild"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/v1.Response"
                        }
                    }
                }
            }
        },
        "/v1/groups-url/{url}": {
            "get": {
                "description": "This API to get group by url.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Groups"
                ],
                "summary": "Get group by url.",
                "parameters": [
                    {
                        "type": "string",
                        "format": "id",
                        "description": "url for get Group",
                        "name": "url",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Groups"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/v1.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/v1.Response"
                        }
                    }
                }
            }
        },
        "/v1/groups/{id}": {
            "get": {
                "description": "This API to get group by id.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Groups"
                ],
                "summary": "Get group by id.",
                "parameters": [
                    {
                        "type": "string",
                        "format": "id",
                        "description": "id for get Group",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Groups"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/v1.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/v1.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/v1.Response"
                        }
                    }
                }
            },
            "put": {
                "description": "This API to update a group.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Groups"
                ],
                "summary": "Update a group.",
                "parameters": [
                    {
                        "type": "string",
                        "format": "id",
                        "description": "id for update Group",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "x-order": "11",
                        "name": "seoDescription",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "x-order": "12",
                        "name": "seoKeywords",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "x-order": "13",
                        "name": "seoText",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "x-order": "14",
                        "name": "seoTitle",
                        "in": "formData"
                    },
                    {
                        "type": "boolean",
                        "x-order": "15",
                        "name": "enabled",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "x-order": "2",
                        "name": "url",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "x-order": "3",
                        "name": "nameRu",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "x-order": "3",
                        "name": "nameUz",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "x-order": "4",
                        "name": "nameEn",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "x-order": "5",
                        "name": "descriptionUz",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "x-order": "6",
                        "name": "descriptionRu",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "x-order": "7",
                        "name": "descriptionEn",
                        "in": "formData"
                    },
                    {
                        "type": "integer",
                        "x-order": "8",
                        "name": "position",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "x-order": "9",
                        "name": "parentGroup",
                        "in": "formData"
                    },
                    {
                        "type": "file",
                        "description": "file_image",
                        "name": "file_image",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Groups"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/v1.Response"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/v1.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/v1.Response"
                        }
                    }
                }
            },
            "delete": {
                "description": "This API to delete a group.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Groups"
                ],
                "summary": "Delete a group.",
                "parameters": [
                    {
                        "type": "string",
                        "format": "id",
                        "description": "id for delete Group",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.Groups": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string",
                    "x-order": "1"
                },
                "image": {
                    "type": "string",
                    "x-order": "10"
                },
                "seoDescription": {
                    "type": "string",
                    "x-order": "11"
                },
                "seoKeywords": {
                    "type": "string",
                    "x-order": "12"
                },
                "seoText": {
                    "type": "string",
                    "x-order": "13"
                },
                "seoTitle": {
                    "type": "string",
                    "x-order": "14"
                },
                "enabled": {
                    "type": "boolean",
                    "x-order": "15"
                },
                "createdAt": {
                    "type": "string",
                    "x-order": "16"
                },
                "updatedAt": {
                    "type": "string",
                    "x-order": "17"
                },
                "deletedAt": {
                    "type": "string",
                    "x-order": "18"
                },
                "url": {
                    "type": "string",
                    "x-order": "2"
                },
                "nameUz": {
                    "type": "string",
                    "x-order": "3"
                },
                "nameRu": {
                    "type": "string",
                    "x-order": "3"
                },
                "nameEn": {
                    "type": "string",
                    "x-order": "4"
                },
                "descriptionUz": {
                    "type": "string",
                    "x-order": "5"
                },
                "descriptionRu": {
                    "type": "string",
                    "x-order": "6"
                },
                "descriptionEn": {
                    "type": "string",
                    "x-order": "7"
                },
                "position": {
                    "type": "integer",
                    "x-order": "8"
                },
                "parentGroup": {
                    "type": "string",
                    "x-order": "9"
                }
            }
        },
        "models.GroupsWithChild": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string",
                    "x-order": "1"
                },
                "image": {
                    "type": "string",
                    "x-order": "10"
                },
                "seoDescription": {
                    "type": "string",
                    "x-order": "11"
                },
                "seoKeywords": {
                    "type": "string",
                    "x-order": "12"
                },
                "seoText": {
                    "type": "string",
                    "x-order": "13"
                },
                "seoTitle": {
                    "type": "string",
                    "x-order": "14"
                },
                "enabled": {
                    "type": "boolean",
                    "x-order": "15"
                },
                "createdAt": {
                    "type": "string",
                    "x-order": "16"
                },
                "updatedAt": {
                    "type": "string",
                    "x-order": "17"
                },
                "deletedAt": {
                    "type": "string",
                    "x-order": "18"
                },
                "url": {
                    "type": "string",
                    "x-order": "2"
                },
                "child": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Groups"
                    },
                    "x-order": "2"
                },
                "nameRu": {
                    "type": "string",
                    "x-order": "3"
                },
                "nameUz": {
                    "type": "string",
                    "x-order": "3"
                },
                "nameEn": {
                    "type": "string",
                    "x-order": "4"
                },
                "descriptionUz": {
                    "type": "string",
                    "x-order": "5"
                },
                "descriptionRu": {
                    "type": "string",
                    "x-order": "6"
                },
                "descriptionEn": {
                    "type": "string",
                    "x-order": "7"
                },
                "position": {
                    "type": "integer",
                    "x-order": "8"
                },
                "parentGroup": {
                    "type": "string",
                    "x-order": "9"
                }
            }
        },
...
        "v1.Response": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        }
    }
}